Add relative orbits

+++ helios/src/main.cpp
@@ -173,7 +173,7 @@ void mouse_cursor_callback(GLFWwindow* window, double pos_x, double pos_y)
 
 void mouse_scroll_callback(GLFWwindow* window, double offset_x, double offset_y)
 {
-	camera.speed += offset_y * 10.0f;
+	camera.speed += (float)offset_y * 10.0f;
 }
 
 void framebuffer_size_callback(GLFWwindow* window, int width, int height)
